chore(auth): refactor AbstractRoleService method names and add new criteria validation
