+++ src/Services/Roles/OnRequestRoleService.php
@@ -2,14 +2,9 @@
 
 namespace Seatplus\Auth\Services\Roles;
 
-use Illuminate\Validation\Rule;
-use Illuminate\Validation\ValidationException;
 use Seatplus\Auth\Enums\RoleMembershipStatus;
 use Seatplus\Auth\Enums\RoleType;
-use Seatplus\Auth\Models\AccessControl\RoleMembership;
 use Seatplus\Auth\Models\User;
-use Seatplus\Eveapi\Models\Alliance\AllianceInfo;
-use Seatplus\Eveapi\Models\Corporation\CorporationInfo;
 
 class OnRequestRoleService extends AbstractRoleService implements RoleServiceInterface
 {
@@ -19,24 +14,7 @@ class OnRequestRoleService extends AbstractRoleService implements RoleServiceInt
      */
     public function addCriteriaForRoleApplication(array $entities): void
     {
-        $this->validate($entities, ['corporation', 'alliance']);
-
-        $this->setRoleType(RoleType::ON_REQUEST);
-
-        $this->resetCriteria();
-
-        foreach ($entities as $entity) {
-
-            $entity_type = match ($entity[1]) {
-                'corporation' => CorporationInfo::class,
-                'alliance' => AllianceInfo::class,
-            };
-
-            $this->setRoleMembership(
-                entity_id: $entity[0],
-                entity_type: $entity_type
-            );
-        }
+        $this->addCriteria($entities, RoleType::ON_REQUEST);
 
         $this->syncMembers();
     }
@@ -78,19 +56,6 @@ class OnRequestRoleService extends AbstractRoleService implements RoleServiceInt
         );
     }
 
-    /**
-     * @throws \Throwable
-     */
-    private function validate(array $entities, array $entity_types): void
-    {
-        $validator = validator($entities, [
-            '*.0' => 'required|integer',
-            '*.1' => ['required', 'string', Rule::in($entity_types)],
-        ]);
-
-        throw_if($validator->fails(), ValidationException::withMessages($validator->errors()->toArray()));
-    }
-
     public function syncMembers(): void
     {
         // remove all members that are not within the criteria
@@ -98,13 +63,5 @@ class OnRequestRoleService extends AbstractRoleService implements RoleServiceInt
 
         // update the status of the members based on the user compliance
         $this->updateMemberStatusBasedOnUserCompliance();
-    }
-
-    private function resetCriteria(): void
-    {
-        RoleMembership::query()
-            ->where('role_id', $this->role->id)
-            ->whereIn('entity_type', [CorporationInfo::class, AllianceInfo::class])
-            ->delete();
     }
 }
